updates to covariance file handling
